feat: separate picsellia contexts
