+++ src/models/contexts/processing/test_picsellia_processing_context.py
@@ -1,5 +1,6 @@
 from typing import Any, Dict, Optional
-
+from src.models.contexts.common.picsellia_context import PicselliaContext
+from picsellia.types.enums import ProcessingType
 from picsellia import DatasetVersion, ModelVersion
 from picsellia.types.enums import ProcessingType
